fix(broker): avoid deployment distribution on reprocessing
* don't distribute the deployment on reprocessing because it may be already reprocessed
* instead, we wait until the reprocessing is done and distribute the pending deployments again
